refactor: group entities together in database module
