+++ apps/backend/src/database/entities/subject.entity.ts
@@ -1,0 +1,3 @@
+import { Base } from './base.entity';
+
+export class Subject extends Base {}
