+++ apps/backend/src/permission/permission.controller.ts
@@ -2,8 +2,8 @@ import { Body, Controller, Delete, Post } from '@nestjs/common';
 import { PermissionService } from './permission.service';
 import { UseJwtAuth } from '@auth/decorators';
 import { GrantPermissionDto } from './dtos/grant-permission.dto';
-import { User } from '@user/user.entity';
 import { RevokePermissionDto } from './dtos/revoke-permission.dto';
+import { User } from '@database/entities';
 
 @Controller('permissions')
 @UseJwtAuth()
