+++ apps/backend/src/permission/permisson.module.ts
@@ -1,10 +1,7 @@
 import { Module } from '@nestjs/common';
 import { PermissionService } from './permission.service';
-import { TypeOrmModule } from '@nestjs/typeorm';
-import { Permission } from './permission.entity';
 
 @Module({
-  imports: [TypeOrmModule.forFeature([Permission])],
   providers: [PermissionService],
 })
 export class PermissionModule {}
