--- apps/backend/src/permission/subject.entity.ts
@@ -1,5 +1,0 @@
-import { Base } from '@common/base.entity';
-
-export class Subject extends Base {
-  public static actions: Record<string, string>;
-}
