+++ apps/backend/src/project/project.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { Project } from './project.entity';
+import { Project } from '../database/entities/project.entity';
 import { ProjectController } from './project.controller';
 import { ProjectService } from './project.service';
